Move circleci/graphql fork into this repo as a package
Rationale: We want to do some finer-grained things with our GraphQL error
messages that currently aren't possible with our GraphQL library, and
there will be less development overhead with this as a package instead
of a separate library.

We can revisit moving this into a separate library down the road, but
for now our API design is unstable and we'll want to be able to make
changes quickly in the CLI to match.