Replace forked graphql client with our own package
This change removes a dependency from our app and replaces it with
only about 100 more lines of code. We'll have to keep the Apache
License as we've modified and redistributed their original code.

Keeping the important parts and exposing the request/response/error
internals to the public API of this package allows us to extend errors
with more fields that our GraphQL gives back.

This also simplifies API code because it means we only have a single
client library to deal with instead of a wrapper on top of a client,
though we're still using http.DefaultClient underneath for `do`.